requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
numpy
pillow
python-multipart